refactor(web): give the search filters a home
The state deciding what a search reads — which sources, document sets,
tags and dates — was spread across three files that were each about
something else. `useFilters` sat in the `lib/hooks.ts` grab bag between
the LLM manager and user groups. `buildFilters`, which freezes that state
into the request body, sat in a 43-line `lib/search/utils.ts` beside an
unrelated `openDocument`. The `Filters` and `TimeRange` types it produces
were in `lib/search/interfaces.ts`, which is mostly chat streaming.

They now sit together in `lib/searchFilters/`. Reading the selection, its
type, and the function that turns one into the other no longer means three
directories.

`FilterManager` is now `SearchFilters` and `useFilters` is
`useSearchFilters`. The old names said how the state is managed rather
than what it holds, and "filter manager" gave no hint that the thing it
manages is search scope.

`useSourcePreferences` comes along. It is not the selection itself, but it
is what writes `selectedSources` and it had no other reason to be in
lib/hooks.ts.

`useTimeRange` does not: it ignored the `initialValue` it accepted, always
started at null, and `useFilters` was its only caller, so it is inlined as
the `useState` it always was. The `useTimeRange` in lib/usage/hooks.ts is
a different hook and is untouched.

Nothing changes about what is sent. `buildFilters` still produces the same
`internal_search_filters`, and the admin document explorer still calls it
for its own endpoint.
