+++ web/src/lib/hooks.ts
@@ -231,151 +231,6 @@ export const useFederatedConnectors = () => {
   };
 };
 
-export const useTimeRange = (initialValue?: DateRangePickerValue) => {
-  return useState<DateRangePickerValue | null>(null);
-};
-
-export interface FilterManager {
-  timeRange: DateRangePickerValue | null;
-  setTimeRange: React.Dispatch<
-    React.SetStateAction<DateRangePickerValue | null>
-  >;
-  selectedSources: SourceMetadata[];
-  setSelectedSources: React.Dispatch<React.SetStateAction<SourceMetadata[]>>;
-  selectedDocumentSets: string[];
-  setSelectedDocumentSets: React.Dispatch<React.SetStateAction<string[]>>;
-  selectedTags: Tag[];
-  setSelectedTags: React.Dispatch<React.SetStateAction<Tag[]>>;
-  getFilterString: () => string;
-  buildFiltersFromQueryString: (
-    filterString: string,
-    availableSources: SourceMetadata[],
-    availableDocumentSets: string[],
-    availableTags: Tag[]
-  ) => void;
-  clearFilters: () => void;
-}
-
-export function useFilters(): FilterManager {
-  const [timeRange, setTimeRange] = useTimeRange();
-  const [selectedSources, setSelectedSources] = useState<SourceMetadata[]>([]);
-  const [selectedDocumentSets, setSelectedDocumentSets] = useState<string[]>(
-    []
-  );
-  const [selectedTags, setSelectedTags] = useState<Tag[]>([]);
-
-  function getFilterString() {
-    const params = new URLSearchParams();
-
-    if (timeRange) {
-      params.set("from", timeRange.from.toISOString());
-      params.set("to", timeRange.to.toISOString());
-    }
-
-    if (selectedSources.length > 0) {
-      const sourcesParam = selectedSources
-        .map((source) => encodeURIComponent(source.internalName))
-        .join(",");
-      params.set("sources", sourcesParam);
-    }
-
-    if (selectedDocumentSets.length > 0) {
-      const docSetsParam = selectedDocumentSets
-        .map((ds) => encodeURIComponent(ds))
-        .join(",");
-      params.set("documentSets", docSetsParam);
-    }
-
-    if (selectedTags.length > 0) {
-      const tagsParam = selectedTags
-        .map((tag) => encodeURIComponent(tag.tag_value))
-        .join(",");
-      params.set("tags", tagsParam);
-    }
-
-    const queryString = params.toString();
-    return queryString ? `&${queryString}` : "";
-  }
-
-  function clearFilters() {
-    setTimeRange(null);
-    setSelectedSources([]);
-    setSelectedDocumentSets([]);
-    setSelectedTags([]);
-  }
-
-  function buildFiltersFromQueryString(
-    filterString: string,
-    availableSources: SourceMetadata[],
-    availableDocumentSets: string[],
-    availableTags: Tag[]
-  ): void {
-    const params = new URLSearchParams(filterString);
-
-    // Parse the "from" parameter as a DateRangePickerValue
-    let newTimeRange: DateRangePickerValue | null = null;
-    const fromParam = params.get("from");
-    const toParam = params.get("to");
-    if (fromParam && toParam) {
-      const fromDate = new Date(fromParam);
-      const toDate = new Date(toParam);
-      if (!isNaN(fromDate.getTime()) && !isNaN(toDate.getTime())) {
-        newTimeRange = { from: fromDate, to: toDate, selectValue: "" };
-      }
-    }
-
-    // Parse sources
-    let newSelectedSources: SourceMetadata[] = [];
-    const sourcesParam = params.get("sources");
-    if (sourcesParam) {
-      const sourceNames = sourcesParam.split(",").map(decodeURIComponent);
-      newSelectedSources = availableSources.filter((source) =>
-        sourceNames.includes(source.internalName)
-      );
-    }
-
-    // Parse document sets
-    let newSelectedDocSets: string[] = [];
-    const docSetsParam = params.get("documentSets");
-    if (docSetsParam) {
-      const docSetNames = docSetsParam.split(",").map(decodeURIComponent);
-      newSelectedDocSets = availableDocumentSets.filter((ds) =>
-        docSetNames.includes(ds)
-      );
-    }
-
-    // Parse tags
-    let newSelectedTags: Tag[] = [];
-    const tagsParam = params.get("tags");
-    if (tagsParam) {
-      const tagValues = tagsParam.split(",").map(decodeURIComponent);
-      newSelectedTags = availableTags.filter((tag) =>
-        tagValues.includes(tag.tag_value)
-      );
-    }
-
-    // Update filter manager's values instead of returning
-    setTimeRange(newTimeRange);
-    setSelectedSources(newSelectedSources);
-    setSelectedDocumentSets(newSelectedDocSets);
-    setSelectedTags(newSelectedTags);
-  }
-
-  return {
-    clearFilters,
-    timeRange,
-    setTimeRange,
-    selectedSources,
-    setSelectedSources,
-    selectedDocumentSets,
-    setSelectedDocumentSets,
-    selectedTags,
-    setSelectedTags,
-    getFilterString,
-    buildFiltersFromQueryString,
-  };
-}
-
 export interface LlmDescriptor {
   name: string;
   provider: string;
@@ -1069,192 +924,3 @@ export const fetchConnectorIndexingStatus = async (
 
   return response.json();
 };
-
-interface UseSourcePreferencesProps {
-  availableSources: ValidSources[];
-  selectedSources: SourceMetadata[];
-  setSelectedSources: (sources: SourceMetadata[]) => void;
-}
-
-interface SourcePreferencesSnapshot {
-  sourcePreferences: Record<string, boolean>; // uniqueKey -> enabled status
-}
-
-const LS_SELECTED_INTERNAL_SEARCH_SOURCES_KEY = "selectedInternalSearchSources";
-
-export function useSourcePreferences({
-  availableSources,
-  selectedSources,
-  setSelectedSources,
-}: UseSourcePreferencesProps) {
-  const [sourcesInitialized, setSourcesInitialized] = useState(false);
-
-  const configuredSources = useMemo(
-    () => getConfiguredSources(availableSources),
-    [availableSources]
-  );
-
-  // Load saved source preferences from localStorage
-  const loadSavedSourcePreferences = (): SourcePreferencesSnapshot | null => {
-    if (typeof window === "undefined") return null;
-    const saved = localStorage.getItem(LS_SELECTED_INTERNAL_SEARCH_SOURCES_KEY);
-    if (!saved) return null;
-    try {
-      const res = JSON.parse(saved);
-
-      // Validate the snapshot structure
-      if (
-        typeof res !== "object" ||
-        res === null ||
-        typeof res.sourcePreferences !== "object" ||
-        res.sourcePreferences === null ||
-        Array.isArray(res.sourcePreferences)
-      ) {
-        return null;
-      }
-
-      // Validate that all values in sourcePreferences are booleans
-      for (const value of Object.values(res.sourcePreferences)) {
-        if (typeof value !== "boolean") {
-          return null;
-        }
-      }
-
-      return res as SourcePreferencesSnapshot;
-    } catch {
-      return null;
-    }
-  };
-
-  const persistSourcePreferencesState = (
-    enabledSources: SourceMetadata[],
-    allKnownSources: SourceMetadata[]
-  ) => {
-    if (typeof window === "undefined") return;
-
-    const enabledKeys = new Set(enabledSources.map((s) => s.uniqueKey));
-
-    const snapshot: SourcePreferencesSnapshot = {
-      sourcePreferences: Object.fromEntries(
-        allKnownSources
-          .filter((src) => src.uniqueKey !== undefined)
-          .map((src) => [src.uniqueKey, enabledKeys.has(src.uniqueKey)])
-      ),
-    };
-
-    localStorage.setItem(
-      LS_SELECTED_INTERNAL_SEARCH_SOURCES_KEY,
-      JSON.stringify(snapshot)
-    );
-  };
-
-  // Initialize sources - load from localStorage or enable all by default
-  useEffect(() => {
-    if (!sourcesInitialized && availableSources.length > 0) {
-      const savedSources = loadSavedSourcePreferences();
-
-      if (savedSources !== null) {
-        // Filter out saved sources that no longer exist
-        const { sourcePreferences } = savedSources;
-
-        // Helper to check if there is a preference for a key
-        const hasPref = (key: string) =>
-          Object.prototype.hasOwnProperty.call(sourcePreferences, key);
-
-        // Get sources with no preference
-        const newSources = configuredSources.filter((source) => {
-          return !hasPref(source.uniqueKey);
-        });
-
-        const enabledSources = configuredSources.filter((source) => {
-          return (
-            hasPref(source.uniqueKey) && sourcePreferences[source.uniqueKey]
-          );
-        });
-
-        // Merge valid saved sources with new sources (enable new sources by default)
-        const mergedSources = [...enabledSources, ...newSources];
-        setSelectedSources(mergedSources);
-
-        // Persist the merged state
-        persistSourcePreferencesState(mergedSources, configuredSources);
-      } else {
-        // First time user or invalid data - enable all sources by default
-        setSelectedSources(configuredSources);
-        persistSourcePreferencesState(configuredSources, configuredSources);
-      }
-      setSourcesInitialized(true);
-    }
-  }, [
-    availableSources,
-    configuredSources,
-    sourcesInitialized,
-    setSelectedSources,
-  ]);
-
-  // Re-initialize when the available source set changes (e.g. switching agents).
-  const prevSourcesKey = useRef(availableSources.join(","));
-  useEffect(() => {
-    const key = availableSources.join(",");
-    if (key !== prevSourcesKey.current) {
-      prevSourcesKey.current = key;
-      setSourcesInitialized(false);
-    }
-  }, [availableSources]);
-
-  const enableSources = (sources: SourceMetadata[]) => {
-    setSelectedSources([...sources]);
-    persistSourcePreferencesState(sources, configuredSources);
-  };
-
-  const enableAllSources = () => {
-    enableSources(configuredSources);
-  };
-
-  const disableAllSources = () => {
-    setSelectedSources([]);
-    persistSourcePreferencesState([], configuredSources);
-  };
-
-  const toggleSource = (sourceUniqueKey: string) => {
-    const configuredSource = configuredSources.find(
-      (s) => s.uniqueKey === sourceUniqueKey
-    );
-    if (!configuredSource) return;
-
-    const isCurrentlySelected = selectedSources.some(
-      (s) => s.uniqueKey === configuredSource.uniqueKey
-    );
-
-    let newSources: SourceMetadata[];
-    if (isCurrentlySelected) {
-      newSources = selectedSources.filter(
-        (s) => s.uniqueKey !== configuredSource.uniqueKey
-      );
-    } else {
-      newSources = [...selectedSources, configuredSource];
-    }
-
-    setSelectedSources(newSources);
-    persistSourcePreferencesState(newSources, configuredSources);
-  };
-
-  const isSourceEnabled = (sourceUniqueKey: string) => {
-    const configuredSource = configuredSources.find(
-      (s) => s.uniqueKey === sourceUniqueKey
-    );
-    if (!configuredSource) return false;
-    return selectedSources.some(
-      (s: SourceMetadata) => s.uniqueKey === configuredSource.uniqueKey
-    );
-  };
-
-  return {
-    sourcesInitialized,
-    enableSources,
-    enableAllSources,
-    disableAllSources,
-    toggleSource,
-    isSourceEnabled,
-  };
-}
