+++ web/src/lib/searchFilters/types.ts
@@ -1,0 +1,37 @@
+import type { Tag } from "@/lib/types";
+import type { SourceMetadata } from "@/lib/search/interfaces";
+import type { DateRangePickerValue } from "@/refresh-components/DateRangePicker";
+
+/** The live selection a user edits: which sources, sets, tags and dates to search. */
+export interface SearchFilters {
+  timeRange: DateRangePickerValue | null;
+  setTimeRange: React.Dispatch<
+    React.SetStateAction<DateRangePickerValue | null>
+  >;
+  selectedSources: SourceMetadata[];
+  setSelectedSources: React.Dispatch<React.SetStateAction<SourceMetadata[]>>;
+  selectedDocumentSets: string[];
+  setSelectedDocumentSets: React.Dispatch<React.SetStateAction<string[]>>;
+  selectedTags: Tag[];
+  setSelectedTags: React.Dispatch<React.SetStateAction<Tag[]>>;
+  getFilterString: () => string;
+  buildFiltersFromQueryString: (
+    filterString: string,
+    availableSources: SourceMetadata[],
+    availableDocumentSets: string[],
+    availableTags: Tag[]
+  ) => void;
+  clearFilters: () => void;
+}
+
+export interface TimeRange {
+  start: Date | string | null;
+  end: Date | string | null;
+}
+
+/** The frozen form of a {@link SearchFilters}, as the backend receives it. */
+export interface Filters {
+  source_type: string[] | null;
+  document_set: string[] | null;
+  updated_at_range: TimeRange | null;
+}
